third_party/WebKit/LayoutTests/webaudio/audioparam-nominal-range-expected.txt

CONSOLE WARNING: line 396: Delay.delayTime.value -1 outside nominal range [0, 1.5]; value will be clamped.
CONSOLE WARNING: line 405: Delay.delayTime.value 4 outside nominal range [0, 1.5]; value will be clamped.
CONSOLE WARNING: line 396: StereoPanner.pan.value -3 outside nominal range [-1, 1]; value will be clamped.
CONSOLE WARNING: line 405: StereoPanner.pan.value 3 outside nominal range [-1, 1]; value will be clamped.
CONSOLE WARNING: line 396: DynamicsCompressor.threshold.value -201 outside nominal range [-100, 0]; value will be clamped.
CONSOLE WARNING: line 405: DynamicsCompressor.threshold.value 1 outside nominal range [-100, 0]; value will be clamped.
CONSOLE WARNING: line 396: DynamicsCompressor.knee.value -1 outside nominal range [0, 40]; value will be clamped.
CONSOLE WARNING: line 405: DynamicsCompressor.knee.value 81 outside nominal range [0, 40]; value will be clamped.
CONSOLE WARNING: line 405: DynamicsCompressor.ratio.value 41 outside nominal range [1, 20]; value will be clamped.
CONSOLE WARNING: line 396: DynamicsCompressor.attack.value -1 outside nominal range [0, 1]; value will be clamped.
CONSOLE WARNING: line 405: DynamicsCompressor.attack.value 3 outside nominal range [0, 1]; value will be clamped.
CONSOLE WARNING: line 396: DynamicsCompressor.release.value -1 outside nominal range [0, 1]; value will be clamped.
CONSOLE WARNING: line 405: DynamicsCompressor.release.value 3 outside nominal range [0, 1]; value will be clamped.
CONSOLE WARNING: line 396: BiquadFilter.frequency.value -1 outside nominal range [0, 24000]; value will be clamped.
CONSOLE WARNING: line 405: BiquadFilter.frequency.value 48001 outside nominal range [0, 24000]; value will be clamped.
CONSOLE WARNING: line 396: Oscillator.frequency.value -48001 outside nominal range [-24000, 24000]; value will be clamped.
CONSOLE WARNING: line 405: Oscillator.frequency.value 48001 outside nominal range [-24000, 24000]; value will be clamped.
CONSOLE WARNING: line 325: Delay.delayTime.setValueAtTime value -1 outside nominal range [0, 1]; value will be clamped.
CONSOLE WARNING: line 326: Delay.delayTime.linearRampToValueAtTime value 2 outside nominal range [0, 1]; value will be clamped.
CONSOLE WARNING: line 327: Delay.delayTime.exponentialRampToValue value 3 outside nominal range [0, 1]; value will be clamped.
CONSOLE WARNING: line 328: Delay.delayTime.setTargetAtTime value -1 outside nominal range [0, 1]; value will be clamped.
CONSOLE WARNING: line 329: Delay.delayTime.setValueCurveAtTime value 1.5 outside nominal range [0, 1]; value will be clamped.
Test AudioParam Nominal Range Values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS GainNode.gain.minValue is equal to -3.4028234663852886e+38.
PASS GainNode.gain.maxValue is equal to 3.4028234663852886e+38.
PASS GainNode.gain.minValue = Math.PI is not equal to 3.141592653589793.
PASS GainNode.gain.minValue is read-only.
PASS GainNode.gain.maxValue = Math.PI is not equal to 3.141592653589793.
PASS GainNode.gain.maxValue is read-only.
PASS Nominal ranges for AudioParam(s) of GainNode are correct.

PASS DelayNode.delayTime.minValue is equal to 0.
PASS DelayNode.delayTime.maxValue is equal to 1.5.
PASS DelayNode.delayTime.minValue = Math.PI is not equal to 3.141592653589793.
PASS DelayNode.delayTime.minValue is read-only.
PASS DelayNode.delayTime.maxValue = Math.PI is not equal to 3.141592653589793.
PASS DelayNode.delayTime.maxValue is read-only.
PASS Set DelayNode.delayTime.value = -1 is equal to 0.
PASS Set DelayNode.delayTime.value = 4 is equal to 1.5.
PASS DelayNode.delayTime was correctly clipped to lie within the nominal range.
PASS Nominal ranges for AudioParam(s) of DelayNode are correct.

PASS AudioBufferSourceNode.playbackRate.minValue is equal to -3.4028234663852886e+38.
PASS AudioBufferSourceNode.playbackRate.maxValue is equal to 3.4028234663852886e+38.
PASS AudioBufferSourceNode.playbackRate.minValue = Math.PI is not equal to 3.141592653589793.
PASS AudioBufferSourceNode.playbackRate.minValue is read-only.
PASS AudioBufferSourceNode.playbackRate.maxValue = Math.PI is not equal to 3.141592653589793.
PASS AudioBufferSourceNode.playbackRate.maxValue is read-only.
PASS AudioBufferSourceNode.detune.minValue is equal to -3.4028234663852886e+38.
PASS AudioBufferSourceNode.detune.maxValue is equal to 3.4028234663852886e+38.
PASS AudioBufferSourceNode.detune.minValue = Math.PI is not equal to 3.141592653589793.
PASS AudioBufferSourceNode.detune.minValue is read-only.
PASS AudioBufferSourceNode.detune.maxValue = Math.PI is not equal to 3.141592653589793.
PASS AudioBufferSourceNode.detune.maxValue is read-only.
PASS Nominal ranges for AudioParam(s) of AudioBufferSourceNode are correct.

PASS StereoPannerNode.pan.minValue is equal to -1.
PASS StereoPannerNode.pan.maxValue is equal to 1.
PASS StereoPannerNode.pan.minValue = Math.PI is not equal to 3.141592653589793.
PASS StereoPannerNode.pan.minValue is read-only.
PASS StereoPannerNode.pan.maxValue = Math.PI is not equal to 3.141592653589793.
PASS StereoPannerNode.pan.maxValue is read-only.
PASS Set StereoPannerNode.pan.value = -3 is equal to -1.
PASS Set StereoPannerNode.pan.value = 3 is equal to 1.
PASS StereoPannerNode.pan was correctly clipped to lie within the nominal range.
PASS Nominal ranges for AudioParam(s) of StereoPannerNode are correct.

PASS DynamicsCompressorNode.threshold.minValue is equal to -100.
PASS DynamicsCompressorNode.threshold.maxValue is equal to 0.
PASS DynamicsCompressorNode.threshold.minValue = Math.PI is not equal to 3.141592653589793.
PASS DynamicsCompressorNode.threshold.minValue is read-only.
PASS DynamicsCompressorNode.threshold.maxValue = Math.PI is not equal to 3.141592653589793.
PASS DynamicsCompressorNode.threshold.maxValue is read-only.
PASS Set DynamicsCompressorNode.threshold.value = -201 is equal to -100.
PASS Set DynamicsCompressorNode.threshold.value = 1 is equal to 0.
PASS DynamicsCompressorNode.threshold was correctly clipped to lie within the nominal range.
PASS DynamicsCompressorNode.knee.minValue is equal to 0.
PASS DynamicsCompressorNode.knee.maxValue is equal to 40.
PASS DynamicsCompressorNode.knee.minValue = Math.PI is not equal to 3.141592653589793.
PASS DynamicsCompressorNode.knee.minValue is read-only.
PASS DynamicsCompressorNode.knee.maxValue = Math.PI is not equal to 3.141592653589793.
PASS DynamicsCompressorNode.knee.maxValue is read-only.
PASS Set DynamicsCompressorNode.knee.value = -1 is equal to 0.
PASS Set DynamicsCompressorNode.knee.value = 81 is equal to 40.
PASS DynamicsCompressorNode.knee was correctly clipped to lie within the nominal range.
PASS DynamicsCompressorNode.ratio.minValue is equal to 1.
PASS DynamicsCompressorNode.ratio.maxValue is equal to 20.
PASS DynamicsCompressorNode.ratio.minValue = Math.PI is not equal to 3.141592653589793.
PASS DynamicsCompressorNode.ratio.minValue is read-only.
PASS DynamicsCompressorNode.ratio.maxValue = Math.PI is not equal to 3.141592653589793.
PASS DynamicsCompressorNode.ratio.maxValue is read-only.
PASS Set DynamicsCompressorNode.ratio.value = 1 is equal to 1.
PASS Set DynamicsCompressorNode.ratio.value = 41 is equal to 20.
PASS DynamicsCompressorNode.ratio was correctly clipped to lie within the nominal range.
PASS DynamicsCompressorNode.attack.minValue is equal to 0.
PASS DynamicsCompressorNode.attack.maxValue is equal to 1.
PASS DynamicsCompressorNode.attack.minValue = Math.PI is not equal to 3.141592653589793.
PASS DynamicsCompressorNode.attack.minValue is read-only.
PASS DynamicsCompressorNode.attack.maxValue = Math.PI is not equal to 3.141592653589793.
PASS DynamicsCompressorNode.attack.maxValue is read-only.
PASS Set DynamicsCompressorNode.attack.value = -1 is equal to 0.
PASS Set DynamicsCompressorNode.attack.value = 3 is equal to 1.
PASS DynamicsCompressorNode.attack was correctly clipped to lie within the nominal range.
PASS DynamicsCompressorNode.release.minValue is equal to 0.
PASS DynamicsCompressorNode.release.maxValue is equal to 1.
PASS DynamicsCompressorNode.release.minValue = Math.PI is not equal to 3.141592653589793.
PASS DynamicsCompressorNode.release.minValue is read-only.
PASS DynamicsCompressorNode.release.maxValue = Math.PI is not equal to 3.141592653589793.
PASS DynamicsCompressorNode.release.maxValue is read-only.
PASS Set DynamicsCompressorNode.release.value = -1 is equal to 0.
PASS Set DynamicsCompressorNode.release.value = 3 is equal to 1.
PASS DynamicsCompressorNode.release was correctly clipped to lie within the nominal range.
PASS Nominal ranges for AudioParam(s) of DynamicsCompressorNode are correct.

PASS BiquadFilterNode.frequency.minValue is equal to 0.
PASS BiquadFilterNode.frequency.maxValue is equal to 24000.
PASS BiquadFilterNode.frequency.minValue = Math.PI is not equal to 3.141592653589793.
PASS BiquadFilterNode.frequency.minValue is read-only.
PASS BiquadFilterNode.frequency.maxValue = Math.PI is not equal to 3.141592653589793.
PASS BiquadFilterNode.frequency.maxValue is read-only.
PASS Set BiquadFilterNode.frequency.value = -1 is equal to 0.
PASS Set BiquadFilterNode.frequency.value = 48001 is equal to 24000.
PASS BiquadFilterNode.frequency was correctly clipped to lie within the nominal range.
PASS BiquadFilterNode.detune.minValue is equal to -3.4028234663852886e+38.
PASS BiquadFilterNode.detune.maxValue is equal to 3.4028234663852886e+38.
PASS BiquadFilterNode.detune.minValue = Math.PI is not equal to 3.141592653589793.
PASS BiquadFilterNode.detune.minValue is read-only.
PASS BiquadFilterNode.detune.maxValue = Math.PI is not equal to 3.141592653589793.
PASS BiquadFilterNode.detune.maxValue is read-only.
PASS BiquadFilterNode.Q.minValue is equal to -3.4028234663852886e+38.
PASS BiquadFilterNode.Q.maxValue is equal to 3.4028234663852886e+38.
PASS BiquadFilterNode.Q.minValue = Math.PI is not equal to 3.141592653589793.
PASS BiquadFilterNode.Q.minValue is read-only.
PASS BiquadFilterNode.Q.maxValue = Math.PI is not equal to 3.141592653589793.
PASS BiquadFilterNode.Q.maxValue is read-only.
PASS BiquadFilterNode.gain.minValue is equal to -3.4028234663852886e+38.
PASS BiquadFilterNode.gain.maxValue is equal to 3.4028234663852886e+38.
PASS BiquadFilterNode.gain.minValue = Math.PI is not equal to 3.141592653589793.
PASS BiquadFilterNode.gain.minValue is read-only.
PASS BiquadFilterNode.gain.maxValue = Math.PI is not equal to 3.141592653589793.
PASS BiquadFilterNode.gain.maxValue is read-only.
PASS Nominal ranges for AudioParam(s) of BiquadFilterNode are correct.

PASS OscillatorNode.frequency.minValue is equal to -24000.
PASS OscillatorNode.frequency.maxValue is equal to 24000.
PASS OscillatorNode.frequency.minValue = Math.PI is not equal to 3.141592653589793.
PASS OscillatorNode.frequency.minValue is read-only.
PASS OscillatorNode.frequency.maxValue = Math.PI is not equal to 3.141592653589793.
PASS OscillatorNode.frequency.maxValue is read-only.
PASS Set OscillatorNode.frequency.value = -48001 is equal to -24000.
PASS Set OscillatorNode.frequency.value = 48001 is equal to 24000.
PASS OscillatorNode.frequency was correctly clipped to lie within the nominal range.
PASS OscillatorNode.detune.minValue is equal to -3.4028234663852886e+38.
PASS OscillatorNode.detune.maxValue is equal to 3.4028234663852886e+38.
PASS OscillatorNode.detune.minValue = Math.PI is not equal to 3.141592653589793.
PASS OscillatorNode.detune.minValue is read-only.
PASS OscillatorNode.detune.maxValue = Math.PI is not equal to 3.141592653589793.
PASS OscillatorNode.detune.maxValue is read-only.
PASS Nominal ranges for AudioParam(s) of OscillatorNode are correct.

PASS PannerNode.positionX.minValue is equal to -3.4028234663852886e+38.
PASS PannerNode.positionX.maxValue is equal to 3.4028234663852886e+38.
PASS PannerNode.positionX.minValue = Math.PI is not equal to 3.141592653589793.
PASS PannerNode.positionX.minValue is read-only.
PASS PannerNode.positionX.maxValue = Math.PI is not equal to 3.141592653589793.
PASS PannerNode.positionX.maxValue is read-only.
PASS PannerNode.positionY.minValue is equal to -3.4028234663852886e+38.
PASS PannerNode.positionY.maxValue is equal to 3.4028234663852886e+38.
PASS PannerNode.positionY.minValue = Math.PI is not equal to 3.141592653589793.
PASS PannerNode.positionY.minValue is read-only.
PASS PannerNode.positionY.maxValue = Math.PI is not equal to 3.141592653589793.
PASS PannerNode.positionY.maxValue is read-only.
PASS PannerNode.positionZ.minValue is equal to -3.4028234663852886e+38.
PASS PannerNode.positionZ.maxValue is equal to 3.4028234663852886e+38.
PASS PannerNode.positionZ.minValue = Math.PI is not equal to 3.141592653589793.
PASS PannerNode.positionZ.minValue is read-only.
PASS PannerNode.positionZ.maxValue = Math.PI is not equal to 3.141592653589793.
PASS PannerNode.positionZ.maxValue is read-only.
PASS PannerNode.orientationX.minValue is equal to -3.4028234663852886e+38.
PASS PannerNode.orientationX.maxValue is equal to 3.4028234663852886e+38.
PASS PannerNode.orientationX.minValue = Math.PI is not equal to 3.141592653589793.
PASS PannerNode.orientationX.minValue is read-only.
PASS PannerNode.orientationX.maxValue = Math.PI is not equal to 3.141592653589793.
PASS PannerNode.orientationX.maxValue is read-only.
PASS PannerNode.orientationY.minValue is equal to -3.4028234663852886e+38.
PASS PannerNode.orientationY.maxValue is equal to 3.4028234663852886e+38.
PASS PannerNode.orientationY.minValue = Math.PI is not equal to 3.141592653589793.
PASS PannerNode.orientationY.minValue is read-only.
PASS PannerNode.orientationY.maxValue = Math.PI is not equal to 3.141592653589793.
PASS PannerNode.orientationY.maxValue is read-only.
PASS PannerNode.orientationZ.minValue is equal to -3.4028234663852886e+38.
PASS PannerNode.orientationZ.maxValue is equal to 3.4028234663852886e+38.
PASS PannerNode.orientationZ.minValue = Math.PI is not equal to 3.141592653589793.
PASS PannerNode.orientationZ.minValue is read-only.
PASS PannerNode.orientationZ.maxValue = Math.PI is not equal to 3.141592653589793.
PASS PannerNode.orientationZ.maxValue is read-only.
PASS Nominal ranges for AudioParam(s) of PannerNode are correct.

PASS ConstantSourceNode.offset.minValue is equal to -3.4028234663852886e+38.
PASS ConstantSourceNode.offset.maxValue is equal to 3.4028234663852886e+38.
PASS ConstantSourceNode.offset.minValue = Math.PI is not equal to 3.141592653589793.
PASS ConstantSourceNode.offset.minValue is read-only.
PASS ConstantSourceNode.offset.maxValue = Math.PI is not equal to 3.141592653589793.
PASS ConstantSourceNode.offset.maxValue is read-only.
PASS Nominal ranges for AudioParam(s) of ConstantSourceNode are correct.

PASS AudioBuffer has no AudioParams as expected.

PASS IIRFilterNode has no AudioParams as expected.

PASS WaveShaperNode has no AudioParams as expected.

PASS ConvolverNode has no AudioParams as expected.

PASS AnalyserNode has no AudioParams as expected.

PASS ScriptProcessorNode has no AudioParams as expected.

PASS PeriodicWave has no AudioParams as expected.

PASS ChannelSplitterNode has no AudioParams as expected.

PASS ChannelMergerNode has no AudioParams as expected.

PASS MediaElementAudioSourceNode has no AudioParams as expected.

PASS MediaStreamAudioDestinationNode has no AudioParams as expected.

PASS AudioListener.positionX.minValue is equal to -3.4028234663852886e+38.
PASS AudioListener.positionX.maxValue is equal to 3.4028234663852886e+38.
PASS AudioListener.positionX.minValue = Math.PI is not equal to 3.141592653589793.
PASS AudioListener.positionX.minValue is read-only.
PASS AudioListener.positionX.maxValue = Math.PI is not equal to 3.141592653589793.
PASS AudioListener.positionX.maxValue is read-only.
PASS AudioListener.positionY.minValue is equal to -3.4028234663852886e+38.
PASS AudioListener.positionY.maxValue is equal to 3.4028234663852886e+38.
PASS AudioListener.positionY.minValue = Math.PI is not equal to 3.141592653589793.
PASS AudioListener.positionY.minValue is read-only.
PASS AudioListener.positionY.maxValue = Math.PI is not equal to 3.141592653589793.
PASS AudioListener.positionY.maxValue is read-only.
PASS AudioListener.positionZ.minValue is equal to -3.4028234663852886e+38.
PASS AudioListener.positionZ.maxValue is equal to 3.4028234663852886e+38.
PASS AudioListener.positionZ.minValue = Math.PI is not equal to 3.141592653589793.
PASS AudioListener.positionZ.minValue is read-only.
PASS AudioListener.positionZ.maxValue = Math.PI is not equal to 3.141592653589793.
PASS AudioListener.positionZ.maxValue is read-only.
PASS AudioListener.forwardX.minValue is equal to -3.4028234663852886e+38.
PASS AudioListener.forwardX.maxValue is equal to 3.4028234663852886e+38.
PASS AudioListener.forwardX.minValue = Math.PI is not equal to 3.141592653589793.
PASS AudioListener.forwardX.minValue is read-only.
PASS AudioListener.forwardX.maxValue = Math.PI is not equal to 3.141592653589793.
PASS AudioListener.forwardX.maxValue is read-only.
PASS AudioListener.forwardY.minValue is equal to -3.4028234663852886e+38.
PASS AudioListener.forwardY.maxValue is equal to 3.4028234663852886e+38.
PASS AudioListener.forwardY.minValue = Math.PI is not equal to 3.141592653589793.
PASS AudioListener.forwardY.minValue is read-only.
PASS AudioListener.forwardY.maxValue = Math.PI is not equal to 3.141592653589793.
PASS AudioListener.forwardY.maxValue is read-only.
PASS AudioListener.forwardZ.minValue is equal to -3.4028234663852886e+38.
PASS AudioListener.forwardZ.maxValue is equal to 3.4028234663852886e+38.
PASS AudioListener.forwardZ.minValue = Math.PI is not equal to 3.141592653589793.
PASS AudioListener.forwardZ.minValue is read-only.
PASS AudioListener.forwardZ.maxValue = Math.PI is not equal to 3.141592653589793.
PASS AudioListener.forwardZ.maxValue is read-only.
PASS AudioListener.upX.minValue is equal to -3.4028234663852886e+38.
PASS AudioListener.upX.maxValue is equal to 3.4028234663852886e+38.
PASS AudioListener.upX.minValue = Math.PI is not equal to 3.141592653589793.
PASS AudioListener.upX.minValue is read-only.
PASS AudioListener.upX.maxValue = Math.PI is not equal to 3.141592653589793.
PASS AudioListener.upX.maxValue is read-only.
PASS AudioListener.upY.minValue is equal to -3.4028234663852886e+38.
PASS AudioListener.upY.maxValue is equal to 3.4028234663852886e+38.
PASS AudioListener.upY.minValue = Math.PI is not equal to 3.141592653589793.
PASS AudioListener.upY.minValue is read-only.
PASS AudioListener.upY.maxValue = Math.PI is not equal to 3.141592653589793.
PASS AudioListener.upY.maxValue is read-only.
PASS AudioListener.upZ.minValue is equal to -3.4028234663852886e+38.
PASS AudioListener.upZ.maxValue is equal to 3.4028234663852886e+38.
PASS AudioListener.upZ.minValue = Math.PI is not equal to 3.141592653589793.
PASS AudioListener.upZ.minValue is read-only.
PASS AudioListener.upZ.maxValue = Math.PI is not equal to 3.141592653589793.
PASS AudioListener.upZ.maxValue is read-only.
PASS Nominal ranges for AudioParam(s) of AudioListener are correct.

PASS All nodes were tested.

PASS successfullyParsed is true

TEST COMPLETE
